+++ nilearn/decomposition/canica.py
@@ -180,7 +180,7 @@ class CanICA(MultiPCA):
         self.threshold = threshold
         self.n_init = n_init
 
-    def _unmix_components(self):
+    def _unmix_components(self, components):
         """Core function of CanICA than rotate components_ to maximize
         independance"""
         random_state = check_random_state(self.random_state)
@@ -189,7 +189,7 @@ class CanICA(MultiPCA):
         # Note: fastICA is very unstable, hence we use 64bit on it
         results = Parallel(n_jobs=self.n_jobs, verbose=self.verbose)(
             delayed(self._cache(fastica, func_memory_level=2))
-            (self.components_.T.astype(np.float64), whiten=True, fun='cube',
+            (components.astype(np.float64), whiten=True, fun='cube',
              random_state=seed)
             for seed in seeds)
 
@@ -238,6 +238,6 @@ class CanICA(MultiPCA):
             Unmasked data to process
 
         """
-        MultiPCA._raw_fit(self, data)
-        self._unmix_components()
+        components = MultiPCA._raw_fit(self, data)
+        self._unmix_components(components)
         return self